backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pgvector==0.2.4
openai==1.3.5
python-dotenv==1.0.0
tiktoken==0.5.2
aiofiles==23.2.1
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
requests==2.31.0
pyyaml==6.0.1